feat: consolidate submissions into dashboard, remove standalone submissions page
- Move full submissions management (filters, cards, status tracking) to /dashboard
- Profile page now shows only public-facing content
- Delete /submissions page, add 301 redirects to /dashboard
- Update all nav links to point to /dashboard for submissions
